+++ Sources/iOS/LatticeApp.swift
@@ -1,0 +1,12 @@
+import LatticeKit
+import SwiftUI
+
+@main
+struct LatticeApp: App {
+    var body: some Scene {
+        WindowGroup {
+            BoardView()
+                .padding()
+        }
+    }
+}